+++ CHANGELOG.md
@@ -18,6 +18,8 @@
   \- descriptions for stored procedures defined with `#[tarantool::proc]`.
 - `proc::module_path` helper function for getting a path to the dynamically
   linked object file in which the given symbol is defined.
+- `msgpack::ArrayWriter` helper struct for generating msgpack arrays from
+    arbitrary serializable data.
 - `tarantool::network::client` alternative async network client.
 - `tarantool::network::client::reconnect` reconnecting async network client based on `network::client`.
 - `tarantool::network::protocol` sans-io (without transport layer) implementation of Tarantool Binary Protocol.